+++ save-recordings/README.md
@@ -2,9 +2,9 @@
 
 ## Development Environment
 
-- Node => v16.14.0+
-- npm => v8.0.0+
-- react => v18.0.0+ 
+- Node -> v16.14.0+
+- npm -> v8.0.0+
+- react -> v18.0.0+ 
 
 ## Useful commands
 
@@ -29,6 +29,47 @@ amplify add api
 ```powershell
 amplify status
 ```
+
+## React Notes
+
+Definitions
+
+- JSX -> Javascript syntax extension
+- DOM -> Document Object Model (*React uses a virtual DOM, which allows us to update the page without refreshing the page*)
+
+
+### **Components**
+
+Components can be formated as classes or functions.
+
+**Functions**
+```jsx
+function Header(props) {
+    return (
+        <div>
+            <h1>{props.title}</h1>
+        </div>
+    )
+}
+<Header title="My title" />
+```
+**Classes**
+```javascript
+class Header extends React.Component {
+    render() {
+        return (
+            <div>
+                <h1> {this.props.title} </h1>
+            </div>
+        )    
+    }
+}
+```
+
+### **React Hooks**
+
+- **useState** -> Returns a stateful value and a function to update it.
+- **useEffect** -> Perform side effects in function components. (*Make HTTP requests*)
 
 ## GraphQL Notes
