+++ save-recordings/src/components/App.js
@@ -1,0 +1,18 @@
+import logo from '../assets/images/team.jpg';
+import '../assets/styles/App.css';
+
+
+function App() {
+  return (
+    <div className="App">
+      <header className="App-header">
+        <img src={logo} className="App-logo" alt="logo" />
+        <p>
+          <br />Hello World ✌️🥴
+        </p>
+      </header>
+    </div>
+  );
+}
+
+export default App;
