+++ save-recordings/src/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
-import App from './App';
-import Home from './App'
+import './assets/styles/index.css';
+// import App from './components/App';
+import Home from './components/Home'
 import reportWebVitals from './reportWebVitals';
 
 import Amplify from 'aws-amplify';
